Added the code to work with Evolio.
The code used to work with Flanco is the only code remaining to be added before going into searching the whole site and comparing the products from Evolio with the products from other sites.

Also, commented out for now, wrote the basic code for comparing a product from a site with another product from another site. The code needs work. In the end, it will only compare two products if they contain the product title found on Evolio. Will probably use RegEx for this.